% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdbase-package.R
\docType{package}
\name{hdbase-package}
\alias{hdbase}
\alias{hdbase-package}
\title{hdbase: Homodatum bases}
\description{
Vector definitions for abstract data types
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datasketch/hdbase}
  \item Report bugs at \url{https://github.com/datasketch/hdbase/issues}
}

}
\author{
\strong{Maintainer}: Juan Pablo Marin Diaz \email{jpmarindiaz@gmail.com}

}
\keyword{internal}
