% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdBase.R
\name{hdBase}
\alias{hdBase}
\title{Create a hdTable data base}
\usage{
hdBase(
  ts,
  name = NULL,
  description = NULL,
  slug = NULL,
  formats = NULL,
  meta = NULL,
  dic = NULL,
  ...
)
}
\arguments{
\item{name}{a custom name can be added}

\item{slug}{a custom slug can be added. If not, hdTable will try creating one.}

\item{meta}{Custom Metadata can be added}

\item{dic}{a custom variable dictionary can be added. \code{\link[=create_dic]{create_dic()}} can help you with that.}

\item{x}{A data frame}

\item{hdTableType}{The type of hdTable to create}

\item{nam}{a custom description can be added}
}
\value{
A hdTable object
}
\description{
Create a hdBase object from a data frame or a list of data frames.
The main value of a hdBase is the metadata associated to a data base.
\itemize{
\item name: Name for the hdTable data frame, setted on \emph{name} argument
\item description: Description for the hdTable data frame, setted on \emph{description} argument
\item hdTables: A list of hdTables
\item slug: a custom slug can be added to the hdBase
}
}
\examples{
hdTable(mtcars, hdTableType = "Num", name = "MTCars")

}
