% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdbase.R
\name{hdbase_hdtables_types}
\alias{hdbase_hdtables_types}
\title{hdbase_hdtables_types}
\usage{
hdbase_hdtables_types(b)
}
\arguments{
\item{x}{an hdbase object}
}
\value{
returns a list of hdtables from the hdbase
}
\description{
get hdtables from base
}
\examples{
some_base <- hdbase(mtcars)
is_hdbase(some_base)

}
