% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdBase.R
\name{is_hdBase}
\alias{is_hdBase}
\title{is_hdBase}
\usage{
is_hdBase(x)
}
\arguments{
\item{x}{object to be coerced or tested}
}
\value{
returns TRUE or FALSE depending on whether its argument is of type hdBase or not.
}
\description{
test for objects of type "hdBase"
}
\examples{
some_base <- hdBase(mtcars)
is_hdBase(some_base)

}
