% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdbase.R
\name{hdbase}
\alias{hdbase}
\title{Create a hdtable data base}
\usage{
hdbase(
  ts,
  name = NULL,
  description = NULL,
  slug = NULL,
  formats = NULL,
  meta = NULL,
  dic = NULL,
  lazy = FALSE,
  ...
)
}
\arguments{
\item{name}{a custom name can be added}

\item{slug}{a custom slug can be added. If not, hdtable will try creating one.}

\item{meta}{Custom Metadata can be added}

\item{dic}{a custom variable dictionary can be added. \code{\link[=create_dic]{create_dic()}} can help you with that.}

\item{x}{A data frame}

\item{hdtable_type}{The type of hdtable to create}

\item{nam}{a custom description can be added}
}
\value{
A hdtable object
}
\description{
Create a hdbase object from a data frame or a list of data frames.
The main value of a hdbase is the metadata associated to a data base.
\itemize{
\item name: Name for the hdtable data frame, setted on \emph{name} argument
\item description: Description for the hdtable data frame, setted on \emph{description} argument
\item hdtables: A list of hdtables
\item slug: a custom slug can be added to the hdbase
}
}
\examples{
hdtable(mtcars, hdtable_type = "Num", name = "MTCars")

}
