% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdbase.R
\name{is_hdbase}
\alias{is_hdbase}
\title{is_hdbase}
\usage{
is_hdbase(x)
}
\arguments{
\item{x}{object to be coerced or tested}
}
\value{
returns TRUE or FALSE depending on whether its argument is of type hdbase or not.
}
\description{
test for objects of type "hdbase"
}
\examples{
some_base <- hdbase(mtcars)
is_hdbase(some_base)

}
