% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdbase-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{hdtype}
\alias{create_dic}
\alias{hdtable}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{hdtable}{\code{\link[hdtable]{create_dic}}, \code{\link[hdtable:hdtable-package]{hdtable}}}

  \item{hdtype}{\code{\link[hdtype:hdtype-package]{hdtype}}}
}}

